requirements.txt
inky~=1.5.0
Django~=4.1.7
Pillow~=9.4.0
requests~=2.28.2
beautifulsoup4==4.12.0
bs4~=0.0.1
djangorestframework~=3.14.0